Fixed one small, but very inconvenient, bug, preventing multiple implementations of a @Service interface to be usable
This is due to the fact that, as they're alternatives implementations of the @Service, they have no primary interface associated.
And they have no primary component, they generate weird bugs in the discovery of any component.
TO prevent that, I've added a fallback case where, if no primary code component is found, the class itself is used as primary component.