Fixes #213
The "easiest" way I found to have both asciidoctor-kroki and asciidoctor-revealjs usable in architecture documentation without relying upon modification of the client POM is to include those gems in my plugin, and unpack those (with the help of the excellent commons-vfs library) on demand. It works !
For a better version, I would invoke an equivalent of the gem-maven-plugin using MojoExecutor, but it's rather hard (it implies having a build extension running, which is not straightforward).